feat: expose `doctor` from every tool
The compositor-capability probe lived only in `wlr-peek`, so a single-tool
install (e.g. `cargo install wlr-shot`) couldn't produce it. Promote its core to
`wlr_capture::doctor` — the report is just a read of the advertised Wayland globals
plus a string table, so it adds no dependencies — and wire it into every binary: a
`doctor` subcommand on wlr-peek/wlr-shot/wlr-draw, a `--doctor` flag on
wlr-chooser/wlr-switcher. The focus-IPC line stays gated behind the `focus` feature
so lean builds don't gain that dependency. Docs updated to match.
